run argument optimization. edge case handling. documentation.

+++ src/com/bj/marketdata/service/DerivedMarketData.java
@@ -3,9 +3,9 @@ package com.bj.marketdata.service;
 public record DerivedMarketData(
         long timestampMillis,
         String instrument,
-        double baseRate,
-        double spread,
-        double adjustment,
-        double derivedValue
+        long baseRate,
+        long spread,
+        long adjustment,
+        long derivedValue
 ) {
 }
